requirements.txt
flask
pyjwt
pymongo
bcrypt
cryptography
pytest
mongomock